support activelearning select valuable samples
