+++ antgo/framework/helper/models/__init__.py
@@ -2,3 +2,4 @@ from .classification import *
 from .semi import *
 from .detectors import *
 from .nas import *
+from .activelearning import *
