+++ antgo/framework/helper/models/activelearning/__init__.py
@@ -1,0 +1,1 @@
+from .acnet import ACModule
